chore(ui): split components
