+++ ui/src/features/stage/stage-details.tsx
@@ -11,6 +11,7 @@ import { Description } from '../common/description';
 import { ManifestPreview } from '../common/manifest-preview';
 import { useImages } from '../project/pipelines/utils/useImages';
 
+import { FreightHistory } from './freight-history';
 import { Promotions } from './promotions';
 import { RequestedFreight } from './requested-freight';
 import { StageActions } from './stage-actions';
@@ -63,21 +64,17 @@ export const StageDetails = ({ stage }: { stage: Stage }) => {
           </div>
           <Divider style={{ marginTop: '1em' }} />
 
-          <div className='flex flex-col gap-8 flex-1'>
-            <div>
-              <RequestedFreight
-                requestedFreight={stage?.spec?.requestedFreight || []}
-                projectName={projectName}
-                itemStyle={{ width: '250px' }}
-                className='space-y-5'
-                freightHistory={stage?.status?.freightHistory || []}
-                currentActiveFreight={stage?.status?.lastPromotion?.freight?.name}
-              />
-            </div>
+          <div className='flex flex-col gap-8 flex-1 pb-10'>
+            <RequestedFreight
+              requestedFreight={stage?.spec?.requestedFreight || []}
+              projectName={projectName}
+              itemStyle={{ width: '250px' }}
+              className='space-y-5'
+            />
             <Tabs
               className='flex-1'
               defaultActiveKey='1'
-              style={{ minHeight: '500px' }}
+              style={{ minHeight: 'fit-content' }}
               items={[
                 {
                   key: '1',
@@ -101,6 +98,13 @@ export const StageDetails = ({ stage }: { stage: Stage }) => {
                   children: <ManifestPreview object={stage} height='700px' />
                 }
               ]}
+            />
+
+            <FreightHistory
+              requestedFreights={stage?.spec?.requestedFreight || []}
+              freightHistory={stage?.status?.freightHistory}
+              currentActiveFreight={stage?.status?.lastPromotion?.freight?.name}
+              projectName={projectName || ''}
             />
           </div>
         </div>
